Created delegate.h and modified remove function
There is now a delegate.h file that has the purpose of containing the delegate code for both the C and C++ versions. WIP.

The Delegate class' remove function has been modified to allow removing one single instance of a function within the invoke list, as well as maintaining the invocation order.

TODO: Removing currently removes the first appearance of the requested function pointer by looking from the begining of the invoke list. Other delegate implementations like C#'s remove the last instance that was inserted into the invoke list. Reconsider which would be the most appropriate way of doing so.