Added InvokeList class
The invoke list class now contains all the logic for handling the container for the function pointers as well as all the related operations. This way, if the implementation of the internal container used changes, the rest of the Delegate related code remains unaffected.

TODO: Template the InvokeList class to allow a custom container to be used, defaulting to std::vector, without actually hard coding the usage of std::vector.