Added a simple function signature example
This example showcases the way that we can make use of C++'s template function signatures to prevent having to explicitly use asterisks for function pointers, as well as simplifying the process of declaring a custom delegate type with a far more user friendly syntax.

For instance, we can change Delegate<int, int, int> d = add;
to
Delegate<int(int,int)> d = add;

This change is not really necessary, as there is already a macro in place to simplify the process of delegate creation, but embracing C++'s template function signatures could lead to far more readable code.