Added note
Non capturing lambdas can decay to function pointers, which means that the current implementation works just fine until you try to use it in more complex code.

From the point of view of C++'s type system, a capturing lambda is a special type, which means that changing to the function signature mentioned in the previous commit coud potentially solve this issue.