Delegates now make use of C++ function signatures
Modified the delegate class to make use of C++'s function signature in its public interface to allow users to more easily define delegates.

As of now, it has been decided that capturing lambdas and such will not be supported, so as to prevent UB from accessing a lambda that captures a value that has gone out of scope.