Added more C++ delegate examples
example_2 demonstrates singlecast delegates.

example_3 demonstrates the copying and assignment of delegates. It also demonstrates how the built in default C++ copy operation works out of the box to perform a deep copy of the vector without having to make an user defined function. This might need to be changed in the future if the class ever gets more complex and actually requires some custom defined logic for deep copying.