Added a C singlecast delegate example
Singlecast delegates are basically just a function pointer wrapper at the end of the day and are the least interesting feature of this library.

TODO: In future updates, I could update the #define system to actually automatically define some information in case other defines are present. For example, RETURN_TYPE could be set to void automatically if RETURN_TYPE_IS_VOID is 1, DELEGATE_IS_MULTICAST could default to 1, etc...